fix: separate bounded text selection UI
